maxBodySize modified

+++ includes/Server.hpp
@@ -37,7 +37,7 @@ class Server {
 		vec_str	 						_serverName;
 		std::string						_root;
 		map_str							_errorPage;
-		size_t							_maxBodySize;
+		long long						_maxBodySize;
 		std::vector<Location>			_location;
 		int								_serverNb;
 		std::vector<Server>				_servers_g;
@@ -50,7 +50,7 @@ class Server {
 		std::vector<size_t>				getPort() const;
 		vec_str							getServerName() const;
 		map_str							getErrorPage() const;
-		size_t							getMaxBodySize() const;
+		long long						getMaxBodySize() const;
 		std::vector<Location>			getLocation() const;
 		size_t							getServerNb() const;
 
@@ -62,7 +62,7 @@ class Server {
 		void	setPort(std::vector<size_t> port);
 		void	setServerName(vec_str serverName);
 		void	setErrorPage(map_str errorPage);
-		void	setMaxBodySize(size_t maxBodySize);
+		void	setMaxBodySize(long long maxBodySize);
 		void	setMaxBodySize(std::string maxBodySize);
 		void	setLocation(std::vector<Location> location);
 		void	setServerNb(size_t serverNb);
